+++ src/netbox/mod.rs
@@ -14,5 +14,6 @@ pub mod models;
 pub mod pagination;
 pub mod prefix_tree;
 pub mod query;
+pub mod rack_elevation;
 pub mod search;
 pub mod status;